Session24/requirements.txt
PyPDF2==3.0.1
python-docx==1.0.1
typing-extensions==4.9.0
google-generativeai==0.3.2 